apt shouldn't be used for scripting.
Ubuntu's man page says that "The apt(8) commandline is designed as an end-user tool and it may change behavior between versions. While it tries not to break backward compatibility this is not guaranteed either if a change seems beneficial for interactive use."

So there's no harm, and possible benefit. So why not change it to the officially recommended tools?